Backend/requirements.txt
fastapi
uvicorn
pydantic
supabase
python-dotenv
sentence-transformers
joblib
scikit-learn
numpy
xgboost
torch
transformers
sentencepiece
google-api-python-client
google-auth-httplib2
google-auth-oauthlib